fix: extend the cast rule to bool, and stop re-pair's filter defeating --package
Addresses PR #375 review (all three findings confirmed against the SDK/build
before acting):

- fw: `pairing_complete()`'s `bonded` is a `bool`, and cbprintf_cxx.h has no
  `bool` overload either — a bool is an EXACT match for the generic template
  (beating every conversion-requiring candidate), so it took the identical
  4-bytes-out-of-a-1-byte-object path and a successful pairing logged
  "bonded: 59609089" instead of "bonded: 1". Now cast.

- fw/CLAUDE.md: the "enum class is safe by construction" line was WRONG and would
  have let the bug straight back in. Zephyr's arg checker is only `__printf_like`,
  so a mismatch is a -Wformat warning, and this build sets no -Werror (verified:
  neither fw/prj.conf nor fw/CMakeLists.txt) — a scoped enum under %d compiles
  clean and ships the same garbage, while an unscoped packed enum warns not at
  all. Rule widened to every enum, scoped or unscoped, and every bool, with the
  warning explicitly demoted from gate to bonus.

- scripts: re-pair.py's install check filtered on a hardcoded "rgbsunglasses",
  which defeats --package for any renamed/forked/staging id — an installed
  package would be reported as missing, the same misleading-failure class the
  check was added to remove. Filters on PKG itself now (keeping the exact
  "package:<name>" comparison, since `pm list packages` substring-matches), and
  derives the failure-message candidate list from PKG's own vendor prefix so a
  renamed package still gets useful suggestions instead of an empty list.

Verified: proto0 builds clean; twister 660/660 cases across 63/63 configs; app
jest 546/546, tsc clean, lint clean; re-pair.py compiles.

Note on size: appcore moved 85.08% -> 85.23% FLASH and 88.15% -> 88.47% RAM, but
both regions moved by EXACTLY +1408 B, which is the kobject_data signature (that
section is counted in both, and its gperf hash resizes on any layout shift) — not
the cost of one cast. See the kobject_data note in the root CLAUDE.md.

+++ fw/src/bluetooth.cpp
@@ -576,7 +576,12 @@ static void pairing_complete(struct bt_conn *conn, bool bonded) {
 
     bt_addr_le_to_str(bt_conn_get_dst(conn), addr, sizeof(addr));
 
-    LOG_INF("Pairing completed: %s, bonded: %d", addr, bonded);
+    // (int) for the same reason as the enum sites below: cbprintf_cxx.h has explicit
+    // promoting overloads for char/short/etc but none for `bool` either, so a bool is an
+    // EXACT match for the generic template (beating every conversion-requiring candidate)
+    // and takes the identical 4-bytes-out-of-a-1-byte-object path. Uncast, a successful
+    // pairing logged "bonded: 59609089" instead of "bonded: 1".
+    LOG_INF("Pairing completed: %s, bonded: %d", addr, (int)bonded);
 }
 
 static void pairing_failed(struct bt_conn *conn, enum bt_security_err reason) {
